Added test_topo setup  folder and xmls
In the process of implementing the proper getting and setting of landmasks.
Created a struct LandMask containing rcp's to both global and local landmasks.

The broadcast in THCM::getLandMask is probably not necessary since we only
need the global mask on proc 0.

However, this is all non-critical initialization stuff so...